WIP: stubbed in moisture map refinement
