+++ Assets/Scripts/WorldGen/Wind.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public static class Wind
+{
+    public static Vector2 FindOrigin(Vector2 coord)
+    {
+        //TODO:
+        return coord;
+    }
+
+    public static float Velocity(Vector2 coord)
+    {
+        //TODO:
+        return 0.0f;
+    }
+}
